dmとchannelのmessageのAPIでファイルを分ける
